Check nbf and exp claims
This commit adds support for the reserved nbf (not before) and exp
(expires) claims. In addition to a valid signature, the current
time to be within the range expressed in the nbf and exp claims.
Both nbf and exp are optional: If omitted or assigned an invalid
value, the lower or upper time boundary does not apply, respectively.